fix api 1

+++ src/app/api/articles/delete/route.ts
@@ -8,12 +8,24 @@ export async function DELETE (request: Request){
     const body = await request.json();
 
     try {
+        if (!token) {
+            return NextResponse.json(
+                {
+                    message: "Cookie is undefined or has no value",
+                },
+                {
+                    status: 400,
+                }
+            );
+        }
+
+        const {value} = token
         const { uid } = body
         console.log(`${process.env.NEXT_PUBLIC_API_URL}/api/articles/${uid}`);
         const result = axios.delete(`${process.env.NEXT_PUBLIC_API_URL}/api/articles/${uid}`, {
             headers: {
                 "Content-Type": "application/json",
-                'Authorization': `Bearer ${token.value}`
+                'Authorization': `Bearer ${value}`
             }
         })
         const { data } = await result
@@ -25,7 +37,7 @@ export async function DELETE (request: Request){
     }catch (error) {
         return NextResponse.json(
             {
-                message: error.message,
+                message: 'Delete Article Failed',
             },
             {
                 status: 400,
